Avoid storing same dataset multiple times
Earlier, it was storing identical copies of the same dataset for each
label, even though only the dataset for the age label was different. Now
for the other labels, only a reference to a common dataset is stored

Also only fit the scaler in the pipeline. Earlier, it was fitting the
scaler twice: once when preprocessing the dataset, and again as part of
the pipeline